fix(fw): make fatal error reporting robust
Use bounded formatting for fatal and parser diagnostics, teach window output to handle line endings, and allocate decoded config blobs to their required size.

+++ fw/src/menu/menu.c
@@ -154,7 +154,11 @@ void action_set_options(void* context, options_t* options) {
             ctx->system_state->pet_display_columns = pet_display_columns_40;
             break;
         default:
-            vet(options->columns == 80, "Invalid 'columns:' in config (got %d)", options->columns);
+            vet(
+                options->columns == 80,
+                "Invalid 'columns:' in config (got %lu)",
+                (unsigned long) options->columns
+            );
             ctx->system_state->pet_display_columns = pet_display_columns_80;
             break;
     }
